requirements.txt
gradio
pandas
fire
torch
diffusers
openai
transformers
Pillow